Criação de CRUD para Cliente

+++ src/main/java/com/accenture/academico/repository/AgenciaRepository.java
@@ -4,7 +4,8 @@ import com.accenture.academico.model.Agencia;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 @Repository
-public interface AgenciaRepository extends JpaRepository<Agencia, Integer> {
-    
+public interface AgenciaRepository extends JpaRepository<Agencia, Long> {
+
 }
