+++ src/main/java/com/accenture/academico/repository/ClienteRepository.java
@@ -4,7 +4,8 @@ import com.accenture.academico.model.Cliente;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 @Repository
-public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
-    
+public interface ClienteRepository extends JpaRepository<Cliente, Long> {
+
 }
